Rename mouthful module 'ancillary_blocks' to 'auxiliary'
