+++ pdtable/auxiliary.py
