+++ pdtable/test/test_ancillary_blocks.py
@@ -2,7 +2,7 @@ from io import StringIO
 from pathlib import Path
 from textwrap import dedent
 
-from ..ancillary_blocks import Directive, MetadataBlock
+from ..auxiliary import Directive, MetadataBlock
 from ..demo.directive_handlers import handle_includes
 from ..io.parsers.blocks import parse_blocks
 from ..store import BlockType
